_development/binance-multi-assets-singal-agent/requirements.txt
python-binance
pandas
pandas-ta
numpy<2
python-dotenv
